+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add request-local thread and Fiber cleanup options
+
 ## [0.19.0] - 2026-08-29
 
 - Skip HTTP/2 Ractor pools without SSL bindings